BigInteger で尋常ではない大きな数字に対応

+++ app/src/main/java/jp/kyuuki/kanjichanger/models/KanjiChangeableNumber.kt
@@ -1,11 +1,12 @@
 package jp.kyuuki.kanjichanger.models
 
 import jp.kyuuki.kanjichanger.components.numberToKanji
+import java.math.BigInteger
 
 /**
  * 漢字変換可能数値.
  */
-class KanjiChangeableNumber(val value: Long) {
+class KanjiChangeableNumber(val value: BigInteger) {
     fun getKanji() : String {
         return numberToKanji(value)
     }
